fileItem.js: show full filename if selected
The space available for the filename under an icon is limited, so
very long filenames get truncated. This patch makes that, when an
icon is selected, the filename can extend up to five lines (no
mattering if it extends over the icon located below), returning
to the single line mode when unselected. This behavior is consistent
with other desktops.