extension.js: Don't nullify the current process
After sending a kill message to the current process, the variable
holding the object launcher was nullified. This was an error, because
it is needed to check the return value. Due to this bug, adding or
removing a monitor makes the relaunching of the desktop to fail.

This patch fixes this.