general: allow to launch the desktop app standalone
Until now, the desktop app waited for an UUID during startup, and
always hide the top bar and made the background transparent. This
makes it boring to launch an instance from command line just to
test something.

This patch mandates a new command line argument, -U, to specify that
the app must run in desktop mode (no decorations, transparent
background, wait for an UUID through STDIN, and don't exit).