general: don't launch not-executable files with +x
When a file in the desktop has the executable bit set, doing double
click on it will try to execute it. This isn't a good idea because,
sometimes, picture files, or texts, have accidentally that bit set,
preventing them from being opened with an application as expected.

This patch ensures that only files with some specific MIME types
will be launched, while all the others will be opened with the
desired application, no matter if they have the +x flag set.