fileItem.js: fit as many lines as possible in filename
This patch ensures that, when a file isn't selected, its filename
will show as many lines as possible, given the available size in the
cell for the icon and label, instead of showing only one, as done
until now.

Unfortunately, this requires an ugly hack: start with one line, and
increment the number until the size of the label is bigger than the
maximum allowed, and in that moment, reduce by one the number of
lines and end.