extension.js: fix path in executable when searching old processes
In previous versions, the shebang at ding.js was '/usr/bin/gjs'; but
it was changed to '/usr/bin/env gjs' to avoid setting the specific
location of the interpreter. This change implied that now, the command
line of the desktop process no longer is '/usr/bin/gjs ...', but
'gjs ...', and this must be reflected in the function for killing old
processes, which is what this patch does.