desktopManager.js: takes into account client-side decorations
When GtkWindow gained client-side decoration capabilities, the
coordinates origin stopped being located in the top-left corner of
the inner area, but it can be outside the window.

This patch takes that into account and now connects to the DRAW
signal in the Gtk.Fixed, and adds a Gtk.EventBox to manage the
mouse events, to ensure that the coordinates, under Wayland and
when using the debug mode, are consistent.