extension.js: Separate wayland and X11 paths
Separate the Wayland and X11 paths, so the desktop icons program,
under X11, will do itself the job of keeping its window at the
bottom and removing from the list of windows.

Also create a new class that encapsulates all the code for launching
the desktop process and checking whether a window belongs to it or
not. This allows to reuse part of the code.