desktopManager.js: Remove coordinates when needed
Now the desktop removes the coordinates from a file in two cases:

- when a file is moved from another folder into the desktop: this is
a must because if the file already has defined coordinates, it could
"fight" for them if there is another file already in those coordinates.

- when a file is hidden and the user doesn't show them: in this case,
this is needed because, if the user decides to show the hidden files,
and there is another file in the same coordinates, the hidden file
would "fight" for the coordinates too.