desktopManager.js: don't delete coordinates during DnD
Until now, when a DnD operation started in the desktop and passed
over a nautilus window, the current coordinates were deleted. This is
an incorrect behavior because it can happen that the user decide to
roll back and don't move the files in that window.

Now the coordinates are preserved. This is not a problem, because
if a file is moved again from other folder into the desktop, the
old coordinates are deleted then.