general: icons must forget coordinates when moved out
When moving an icon outside of the desktop, it must "forget" their
coordinates to ensure that, if sent again into the desktop, it won't
fight with another one added after.

Also, although coordinates are deleted when a file is added to the
desktop, when adding several files at once this fails, because it
deletes the coordinates of the first file and triggers a desktop
repaint, where the coordinates of all the other icons are set.

This patch fixes this by ensuring that the coordinates are removed
when a file is moved out of the desktop.