extension.js: don't relaunch when re-enabling fast
When an extension is disabled, all the extensions installed before
it are disabled before, in reverse order, to ensure that callbacks
and hooks are correctly unset, and are enabled again after the
desired extension has been disabled. This process lasts very little
time.

In DING, when it is disabled, the current desktop process is killed,
and when it is enabled, a new process is launched. The problem is
that, if this process is too fast, when the callback for the first
child being killed arrives, the new child has been already launched,
and a new, third child is launched, so the second, the one that
should be the only one, is left orphan and unmanaged, so it will be
shown over all other windows.

This patch fixes this by ensuring that the process that died is the
one that was currently in use, and not an old one, before launching
a new child.