general: fix DnD precission in multimonitor systems
When using multimonitor environments, it is not possible to drop an
icon in the first column of a monitor different from the leftmost
one. This is due to round errors.

This patch fixes it, by calculating the coordinates of the final
grid possition, instead of using the direct destination coordinates
of the DnD operation.