Fixed border effects on multimonitor and zoom modes
When using a multimonitor environment and one monitor has a different
zoom setting, some icons placed at the border are sent to the wrong
screen. This was because the test to determine if an icon fall in an
specific screen didn't take into account the zoom factor, so only
those icons placed in the half upper part (for a zoom of 200%) where
being placed correctly.

This patch fixes it.