Add support for custom icons
Nautilus allows to set a custom icon for a file. This patch
implements it. Unfortunately, there is no way of knowing when
a custom icon has been set or unset, so it is a must to refresh
the desktop with F5 to show any change.

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/50