Rely only in CHANGES_DONE_HINT
The current code refreshes the desktop on any file event from the
desktop. The problem with this is that the CHANGED event is sent
even when the modifications to a file hasn't ended. This can result
in a thumbnail being created before the picture has been fully
stored, resulting in a half-backed thumbnail.

This patch ignores the CHANGED event to avoid this; thus only
the CHANGES_DONE_HINT event will trigger a desktop refresh and
a recreation of thumbnails.