Re-read the desktop on error
When reading the desktop and an error happens, re-read it. Before,
it exited, which caused the extension to reload it, but when editing
files in the desktop (like with Libreoffice) a lot of them occurred
because a file was removed, another one created... and the desktop
flickered.

This patch removes that flicker.