Better detection of origin is destination in DnD
The last patch added some code to detect when the user drops a
file or folder over itself (which means that (s)he has changed
its mind and want to abort the DnD operation) which was quite
ugly and hacky.

This new patch reworks the concept and makes it much more
elegant.