Removed the need to keep the original coordinates
In local DnD (from desktop to desktop), a trick was used to know
the origin coordinates, which was to send it as part of the DnD
data. But now that we store the original fileItem object that
starts the DnD operation, this is needed no more.

This patch removes that from the code.