Update for Gnome Shell 3.38
This patch takes advantage of the new API available in Gnome
Shell 3.38 for detecting if a window belongs to a process
launched from an extension. This allows to remove the ugly
hack of sending an UUID through STDIN and put it in the window's
title to ensure that a window was from a trusted program.

It also makes use of the available call to remove a window from
the window list, which finally fixes having an icon in the dash
for each desktop.

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/45
Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/48