RETURN works when creating folder
When creating a folder, pressing RETURN does the same than
the OK button. Unfortunately, it is managed also when the OK
button is disabled (like when there is already a folder with
the same name).

This patch fixes that, and the same problem during renaming.

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/79