Select a new file created from templates
When a file is created from a template, it has a generic name,
so it is common to rename it immediately after creating it.
This patch makes any new file created, to be "selected", thus
just pressing F2 will trigger the RENAME option.

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/82