Don't maximize a window when closing popup menu
Every time a window is destroyed, the topmost window receives
the focus. Unfortunately, this means that if the topmost window
is minimized and the window closed is a popup menu or a tooltip
from DING, it will unminimze the window.

This patch fixes it.

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/83