dbusUtils: Use org.gnome.Nautilus.FileOperations2 API
It supports platform data that allows the callers to pass more data such
as the event timeout that can will be used to raise the eventual dialogs
such as in the empty-trash case.

dbusUtils: Try to get the parent handle from the event and pass it to nautilus

desktopManager: Use NautilusFileOperations2 to delete files recursively

This feature is now provided directly by nautilus, so no need to
re-implement it in the extension, so that can be handled natively,
removing lots of code that was added to mimic the native behavior.

This also fixes the ability of permanently removing folders.