Don't repaint the whole desktop if metadata is updated
When a file's metadata changes (like when the user sets a custom
icon for a file, or restores the default one), the full desktop
is repainted. This has a problem: when one or more files are
moved in the desktop, their coordinates (stored in metadata)
change, and it forces a full desktop repaint. This has the
problem that the current selection is erased, and thus all the
selected files got unselected.

This MR fixes this by calling a method in the file whose metadata
changed, and letting it to update whatever it has to update.