Fix default location in network mounts
When network mount desktop icons are clicked, a Nautilus
window will open. However, Nautilus opens the root folder of
the mount, and not the default location of the mount, which
may be different in case gio mount was done with some
subdirectories.

The following patch fixes this by replacing get_root() calls
with get_default_location(). However, this is not a complete
fix, because the folder name will no longer show the
"$FOLDER on $SERVER" message, instead it shows only the
network folder basename. This is due to some problem that
the default_location (at least on Gnome 3.30
(Debian Buster)) lacks "/" in the end in case subdirectories
are involved.

Patch sent by Juha Erkkilä

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/138