Use full surface as DnD destination
After the change to make the selection smaller for the icons,
the destination surface when doing a DnD were also reduced.
This is wrong because you can try to drop an icon in the same
grid zone where there is another file/folder, and it will be
put in an adjacent free grid, when it should be dropped in the
file/folder inside that grid.

This MR fixes this by setting the DnD destination to the whole
icon grid, but only when the "Show DnD destination" option is
enabled.