Allows to DnD in the place of a moving icon
When doing DnD, if the uset drops a group of icons over one of
the icons being moved, the desktop manages it as if the user
wanted to open the files with that icon, instead of assume
that the user just wants to move the icons into that place.

This MR fixes this.

Fix https://gitlab.com/rastersoft/desktop-icons-ng/-/issues/166