Destroy fileItem when removed from grid
Until now, when the desktop was refreshed, the fileItem objects
were just removed from the grid, but any pending operation or
any connected signal was kept until its GUI element signaled
its destruction.

This is a waste of resources, and even can result in spurious
signals being managed after the object has been removed.

This patch cancels all those elements as soon as the fileItem
is removed from the grid.