Several cancellable fixes in fileItem
If the trash content is being read and a changed event is
triggered, now the read is cancelled.

When loading the bytes of a thumbnail, don't check if the
cancellable has been cancelled. That's already managed by
load_bytes_finish(), throwing a CANCELLED exception.

In the callbacks for async functions, always set first the
cancellable to NULL before reading the final status. This
prevents calling cancel() in old cancellables.