extension: Ignore workarea changes if it's just switching to fullscreen
If the monitor is covered by a fullscreen window then we don't need to
react to that workarea change.

Not reloading has the added benefit of not briefly stealing focus from
the window that's going fullscreen. Some fullscreen-only apps like games
would take that to mean that they should minimize themselves. So they
would never be able to stay fullscreen without instantly minimizing again.

Now that shouldn't happen, and there's much less flicker.

Fixes: https://launchpad.net/bugs/1946307